jau::cfmt: Fix append_integral* dest_maxlen clipping (AS-Safe)
Apply dest_maxlen clipping on all layout length variables,
ensuring no 'breakout' of allowed free range within destination string.

Fixed functions:
- append_integral
- append_integral_simple

This fix ensures AS-Safety.

+++ include/jau/string_cfmt.hpp
@@ -491,7 +491,7 @@ namespace jau::cfmt {
     }
 
     namespace impl {
-        inline constexpr const size_t char32buf_maxlen = 32;
+        inline constexpr const uint32_t char32buf_maxlen = 32;
         inline constexpr const uint32_t default_float_precision = 6;
         inline constexpr const double_t max_append_float = (double_t)1e9;
